requirements.txt
python_json_logger==2.0.7
requests==2.31.0
python-telegram-bot==13.1